% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pertruber.R
\name{adaptive_noise}
\alias{adaptive_noise}
\title{Function factory to apply white noise to a vector proportioanl to the spread
of the data}
\usage{
adaptive_noise(sd.ratio = 1/10)
}
\arguments{
\item{sd.ratio}{the level of noise to apply relative to the vectors standard
deviation.}
}
\value{
a function
}
\description{
Function factory to apply white noise to a vector proportioanl to the spread
of the data
}
\examples{

f <- adaptive_noise(0.2)
f(1:10)

}
