% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_deident.R
\name{apply_deident}
\alias{apply_deident}
\title{Apply a 'deident' pipeline}
\usage{
apply_deident(object, deident, ...)
}
\arguments{
\item{object}{The data to be deidentified}

\item{deident}{A deidentification pipeline to be used.}

\item{...}{Terms to be passed to other methods}
}
\description{
Applies a pipeline as defined by \code{deident} to a data frame. tibble, or file.
}
