% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deident.R
\name{apply_to_data_frame}
\alias{apply_to_data_frame}
\title{Apply a 'deident' pipeline to a new data frame}
\usage{
apply_to_data_frame(data, transformer, ...)
}
\arguments{
\item{data}{The data set to be converted}

\item{transformer}{The pipeline to be used}

\item{...}{To be passed on to other methods}
}
\description{
Apply a 'deident' pipeline to a new data frame
}
