% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{babyname_sample}
\alias{babyname_sample}
\title{Babyname prevalence for 2017 and 2018}
\format{
A data frame with 65,448 rows and 5 columns:
\describe{
  \item{year}{Calendar year considered}
  \item{sex}{Assigned at birth gender - M: male, F: female}
  \item{name}{Popular baby name in given year}
  \item{n}{Frequency to which `name` was used in `year`}
  \item{prop}{Proportion of all baby's born in `year` with `name`}

}
}
\usage{
babyname_sample
}
\description{
Subsample of the `babynames` data set found in
\href{https://github.com/hadley/babynames}{`babynames`}
supplied for testing the package.
}
\keyword{datasets}
