% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_deident.R
\name{BaseDeident}
\alias{BaseDeident}
\title{Base class for all De-identifier classes}
\description{
Base class for all De-identifier classes

Base class for all De-identifier classes
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{method}}{Function to call for data transform.
Create new Deidentifier object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BaseDeident-new}{\code{BaseDeident$new()}}
\item \href{#method-BaseDeident-set_method}{\code{BaseDeident$set_method()}}
\item \href{#method-BaseDeident-persist}{\code{BaseDeident$persist()}}
\item \href{#method-BaseDeident-transform}{\code{BaseDeident$transform()}}
\item \href{#method-BaseDeident-mutate}{\code{BaseDeident$mutate()}}
\item \href{#method-BaseDeident-group_mutate}{\code{BaseDeident$group_mutate()}}
\item \href{#method-BaseDeident-group_and_mutate}{\code{BaseDeident$group_and_mutate()}}
\item \href{#method-BaseDeident-serialize}{\code{BaseDeident$serialize()}}
\item \href{#method-BaseDeident-str}{\code{BaseDeident$str()}}
\item \href{#method-BaseDeident-clone}{\code{BaseDeident$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseDeident-new"></a>}}
\if{latex}{\out{\hypertarget{method-BaseDeident-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseDeident$new(method = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{(optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
`Encrypter`
Setter for 'method' field
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseDeident-set_method"></a>}}
\if{latex}{\out{\hypertarget{method-BaseDeident-set_method}{}}}
\subsection{Method \code{set_method()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseDeident$set_method(method)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{New function to be used as the method.
Save 'Deidentifier' to serialized object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseDeident-persist"></a>}}
\if{latex}{\out{\hypertarget{method-BaseDeident-persist}{}}}
\subsection{Method \code{persist()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseDeident$persist(location)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{location}}{File path to save to.
Apply 'method' to a vector of values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseDeident-transform"></a>}}
\if{latex}{\out{\hypertarget{method-BaseDeident-transform}{}}}
\subsection{Method \code{transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseDeident$transform(keys, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{Vector of values to be processed}

\item{\code{...}}{Values to be concatenated to keys
Apply 'method' to variables in a data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseDeident-mutate"></a>}}
\if{latex}{\out{\hypertarget{method-BaseDeident-mutate}{}}}
\subsection{Method \code{mutate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseDeident$mutate(data, ..., force = T)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A data frame to be manipulated}

\item{\code{...}}{Symbol representations of variables in 'data'
Apply 'mutate' method to an aggregated data frame.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseDeident-group_mutate"></a>}}
\if{latex}{\out{\hypertarget{method-BaseDeident-group_mutate}{}}}
\subsection{Method \code{group_mutate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseDeident$group_mutate(grouped_data, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{grouped_data}}{a 'grouped_df' object}

\item{\code{...}}{Symbol representations of variables in 'data'
Aggregate a data frame and apply 'mutate' to each.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseDeident-group_and_mutate"></a>}}
\if{latex}{\out{\hypertarget{method-BaseDeident-group_and_mutate}{}}}
\subsection{Method \code{group_and_mutate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseDeident$group_and_mutate(data, grp_cols, mutate_cols)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A data frame to be manipulated}

\item{\code{grp_cols}}{Vector of variables in 'data' to group on.}

\item{\code{mutate_cols}}{Vector of variables in 'data' to transform.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseDeident-serialize"></a>}}
\if{latex}{\out{\hypertarget{method-BaseDeident-serialize}{}}}
\subsection{Method \code{serialize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseDeident$serialize(type = class(self), on_init = list(), dots = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseDeident-str"></a>}}
\if{latex}{\out{\hypertarget{method-BaseDeident-str}{}}}
\subsection{Method \code{str()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseDeident$str(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseDeident-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BaseDeident-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseDeident$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
