% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blurer.R
\name{Blurer}
\alias{Blurer}
\title{Deidentifier class for applying 'blur' transform}
\arguments{
\item{blur}{Look-up list to define aggregation.}

\item{keys}{Vector of values to be processed}

\item{...}{Values to be concatenated to keys}
}
\value{
\code{Blurer}
Apply blur to a vector of values
}
\description{
Convert \code{self} to a list.
}
\details{
'Bluring' refers to aggregation of data e.g. converting city to country,
or post code to IMD.  The level of blurring is defined by the list given
at initialization which maps key to value e.g.
list(London = "England", Paris = "France").
}
\section{Fields}{

\describe{
\item{\code{blur}}{List of aggregations to be applied.
Create new Blurer object}
}}

