% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_blur.R
\name{category_blur}
\alias{category_blur}
\title{Utility for producing 'blur'}
\usage{
category_blur(vec, ...)
}
\arguments{
\item{vec}{The vector of values to be used}

\item{...}{\code{Replacement} = \code{RegexPattern} pairs of arguments}
}
\description{
Utility for producing 'blur'
}
