% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_deident.R
\name{create_deident}
\alias{create_deident}
\title{Create a deident pipeline}
\usage{
create_deident(method, ...)
}
\arguments{
\item{method}{A deidentifier to initialize.}

\item{...}{list of variables to be deidentifier. NB: key word arguments will
be passed to method at initialization.}
}
\description{
Create a deident pipeline
}
