% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deident_job.R
\name{deident_job_from_folder}
\alias{deident_job_from_folder}
\title{Apply a pipeline to files on disk.}
\usage{
deident_job_from_folder(
  deident_pipeline,
  data_dir,
  result_dir = "Deident_results"
)
}
\arguments{
\item{deident_pipeline}{The deident list to be used.}

\item{data_dir}{a path to the files to be transformed.}

\item{result_dir}{a path to where files are to be saved.}
}
\description{
Apply a deident pipeline to a set of files and save them back to disk
}
