% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_yaml.R
\name{from_yaml}
\alias{from_yaml}
\title{Restore a serialized deident from file}
\usage{
from_yaml(path)
}
\arguments{
\item{path}{Path to serialized deident.}
}
\description{
Restore a serialized deident from file
}
\examples{
\dontrun{
deident <- deident(ShiftsWorked, Pseudonymizer, Employee)
deident$to_yaml("model.yml")

deident.yaml <- from_yaml("model.yml")
deident.yaml$mutate(ShiftsWorked)
}

}
