% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pertruber.R
\name{lognorm_noise}
\alias{lognorm_noise}
\title{Function factory to apply log-normal noise to a vector}
\usage{
lognorm_noise(sd = 0.1)
}
\arguments{
\item{sd}{the standard deviation of noise to apply.}
}
\value{
a function
}
\description{
Function factory to apply log-normal noise to a vector
}
\examples{

f <- lognorm_noise(1)
f(1:10)

}
