% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{new_deident}
\alias{new_deident}
\title{Deidentification API root}
\usage{
new_deident(object, ..., encrypter)
}
\arguments{
\item{object}{Either a \code{data.frame}, \code{tibble}, or existing \code{DeidentList} pipeline.}

\item{...}{variables to be transformed.}
}
\value{
A 'DeidentList' representing the untrained transformation pipeline.
The object contains fields:
\itemize{
\item \code{deident_methods} a list of each step in the pipeline (consisting of \code{variables} and \code{method})
}

and methods:
\itemize{
\item \code{mutate} apply the pipeline to a new data set
\item \code{to_yaml} serialize the pipeline to a '.yml' file
}
}
\description{
A general function for defining a deident function.
}
\keyword{internal}
