% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pertruber.R
\name{Perturber}
\alias{Perturber}
\title{R6 class for deidentification via random noise}
\arguments{
\item{noise}{a single-argument function that applies randomness.}

\item{keys}{Vector of values to be processed}

\item{...}{Values to be concatenated to keys}
}
\description{
A \code{Deident} class dealing with the addition of random noise to a
numeric variable.

Create new Perturber object

Apply noise to a vector of values

Convert \code{self} to a list.

Character representation of the class
}
\section{Fields}{

\describe{
\item{\code{noise.str}}{character representation of \code{noise}}

\item{\code{method}}{random noise function}
}}

\examples{
  pert <- Perturber$new()
  pert$transform(1:10)

}
