% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudonimizer.R
\name{Pseudonymizer}
\alias{Pseudonymizer}
\title{R6 class for deidentification via replacement}
\arguments{
\item{lookup}{a pre-existing name-value pair to define intended psuedonymizations.
Instances of 'name' will be replaced with 'value' on transformation.}

\item{keys}{value to be checked}

\item{...}{values to concatenate to \code{key} and check}

\item{parse_numerics}{True: Force columns to characters.  NB: only
character vectors will be parsed.}
}
\description{
A \code{Deident} class dealing with the (repeatable) random replacement of
string for deidentification.

Create new \code{Pseudonymizer} object

Check if a key exists in \code{lookup}

Check if a key exists in \code{lookup}

Retrieve a value from \code{lookup}

Returns \code{self$lookup} formatted as a tibble

Convert \code{self} to a list.

Apply the deidentifcation method to the supplied \code{keys}
}
\section{Fields}{

\describe{
\item{\code{lookup}}{list of mapping from key-value on transform.}
}}

