% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{set_path_class}
\alias{set_path_class}
\title{read_data <- function(file_path, ...){
#' @export
UseMethod("read_data")
}}
\usage{
set_path_class(file_path)
}
\description{
read_data <- function(file_path, ...){
#' @export
UseMethod("read_data")
}
}
