% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pertruber.R
\name{white_noise}
\alias{white_noise}
\title{Function factory to apply white noise to a vector}
\usage{
white_noise(sd = 0.1)
}
\arguments{
\item{sd}{the standard deviation of noise to apply.}
}
\value{
a function
}
\description{
Function factory to apply white noise to a vector
}
\examples{

f <- white_noise(1)
f(1:10)

}
