% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffler.R
\name{Shuffler}
\alias{Shuffler}
\title{Shuffler class for applying 'shuffling' transform}
\description{
Shuffler class for applying 'shuffling' transform

Shuffler class for applying 'shuffling' transform
}
\section{Super class}{
\code{\link[deident:BaseDeident]{deident::BaseDeident}} -> \code{Shuffler}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Shuffler-new}{\code{Shuffler$new()}}
\item \href{#method-Shuffler-set_limit}{\code{Shuffler$set_limit()}}
\item \href{#method-Shuffler-transform}{\code{Shuffler$transform()}}
\item \href{#method-Shuffler-serialize}{\code{Shuffler$serialize()}}
\item \href{#method-Shuffler-clone}{\code{Shuffler$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="deident" data-topic="BaseDeident" data-id="group_and_mutate"><a href='../../deident/html/BaseDeident.html#method-BaseDeident-group_and_mutate'><code>deident::BaseDeident$group_and_mutate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="deident" data-topic="BaseDeident" data-id="group_mutate"><a href='../../deident/html/BaseDeident.html#method-BaseDeident-group_mutate'><code>deident::BaseDeident$group_mutate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="deident" data-topic="BaseDeident" data-id="mutate"><a href='../../deident/html/BaseDeident.html#method-BaseDeident-mutate'><code>deident::BaseDeident$mutate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="deident" data-topic="BaseDeident" data-id="persist"><a href='../../deident/html/BaseDeident.html#method-BaseDeident-persist'><code>deident::BaseDeident$persist()</code></a></span></li>
<li><span class="pkg-link" data-pkg="deident" data-topic="BaseDeident" data-id="set_method"><a href='../../deident/html/BaseDeident.html#method-BaseDeident-set_method'><code>deident::BaseDeident$set_method()</code></a></span></li>
<li><span class="pkg-link" data-pkg="deident" data-topic="BaseDeident" data-id="str"><a href='../../deident/html/BaseDeident.html#method-BaseDeident-str'><code>deident::BaseDeident$str()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shuffler-new"></a>}}
\if{latex}{\out{\hypertarget{method-Shuffler-new}{}}}
\subsection{Method \code{new()}}{
'Shuffling' refers to the a random sampling of a variable without
replacement e.g. [A, B, C] becoming [B, A, C] but not [A, A, B].
Shuffling will preserve top level metrics (e.g. mean, median, mode) but
removes ordinal properties i.e. correlations and auto-correlations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shuffler$new(method = sample, limit = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{[optional] A function representing the method of re-sampling
to be used.  By default uses exhaustive sampling without replacement.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shuffler-set_limit"></a>}}
\if{latex}{\out{\hypertarget{method-Shuffler-set_limit}{}}}
\subsection{Method \code{set_limit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shuffler$set_limit(limit)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shuffler-transform"></a>}}
\if{latex}{\out{\hypertarget{method-Shuffler-transform}{}}}
\subsection{Method \code{transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shuffler$transform(keys, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shuffler-serialize"></a>}}
\if{latex}{\out{\hypertarget{method-Shuffler-serialize}{}}}
\subsection{Method \code{serialize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shuffler$serialize(dots = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shuffler-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Shuffler-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shuffler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
